backend/pyproject.toml
[project]
name = "flips-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.125.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.40.0",
]
